pyproject.toml
[project]
name = "toggl2jira"
version = "0.0.0"
# Version will be overwritten by pipeline
description = "Replicate toggl time entries to Jira worklogs"
authors = [{ "name" = "Jakob Schlagenhaufer" }, { "name" = "Jan Dittrich" }]
readme = "README.md"
requires-python = ">=3.13"
dynamic = ["dependencies"]

[project.scripts]
toggl2jira = "toggl2jira.cli:main"

[tool.poetry]
requires-poetry = ">=2.0"

[tool.poetry.dependencies]
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
python-dotenv = "^1.0.1"
truststore = "^0.10.1"

[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
isort = "^6.0.1"
pylint = "^3.3.8"

[tool.isort]
profile = "black"
skip_gitignore = true

[tool.pylint]
max-line-length = 120

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
